Steps taken towards functionality of libpng implementation
file IO with libpng working, but only for RGB with 3 channels. Also, the
hiding and recovering process is not yet functional and has to be worked
on further. Code has been somewhat refactored and moved into header and
source files. A help message to the user has been added, which lists the
flags that will be added in a future commit