Improvements made to the cli interface and the encoder function
it appears I was missing a bitshift and as a result I was storing the
least significant bits of the secret image into the cover. The cli
interface is improved and now accepts flags. Some user input that has
been given flags is yet to be utilized in the algorithm ( for example,
the number of bits to replace when combining)