Makefile now has install, uninstall, and allclean targets
install and uninstall create and remove an executable "steganize" in
the /bin directory. As a result, they require root permissions. The
allclean target removes all the files in the program's repository, and
any executables made while using it (the /bin executable will be erased
only if run with root permissions)